split actual importer from demo loader
